add casts to ignore narrowing conversion

+++ src/wmtk/TriMesh.cpp
@@ -116,7 +116,10 @@ void TriMesh::initialize(
 {
     // reserve memory for attributes
 
-    std::vector<long> cap{FF.rows(), VF.rows(), EF.rows()};
+    std::vector<long> cap{
+        static_cast<long>(FF.rows()),
+        static_cast<long>(VF.rows()),
+        static_cast<long>(EF.rows())};
     set_capacities(cap);
     reserve_attributes_to_fit();
